Unify context picker across template launch and create-agent dialogs
Extract shared ContextPicker component from the template launch dialog
and use it in both the create-agent and template launch flows, ensuring
consistent UX for file/link/clipboard attachment across the app.
