+++ apps/web/src/components/app/create-agent-dialog.tsx
@@ -1,5 +1,4 @@
 import {
-  type ChangeEvent,
   type ClipboardEvent,
   type DragEvent,
   type FormEvent,
@@ -10,20 +9,15 @@ import {
   useState,
 } from "react";
 import { useAtom } from "jotai";
-import {
-  Check,
-  ChevronDown,
-  Clipboard,
-  ChevronLeft,
-  FileText,
-  GitBranch,
-  Link2,
-  Plus,
-  Upload,
-  X,
-} from "lucide-react";
+import { Check, ChevronDown, ChevronLeft, GitBranch } from "lucide-react";
 
 import { BranchSelect } from "@/components/app/branch-select";
+import { ContextPicker } from "@/components/app/context-picker";
+import {
+  createClipboardSuggestionFromText,
+  getClipboardFilesFromEvent,
+  startupFileKey,
+} from "@/components/app/create-agent-dialog-clipboard";
 import { PathInput } from "@/components/app/path-input";
 import { ActivityBars } from "@/components/ui/activity-bars";
 import { Button } from "@/components/ui/button";
@@ -42,22 +36,6 @@ import {
   DialogTitle,
 } from "@/components/ui/dialog";
 import { Input } from "@/components/ui/input";
-import {
-  type ClipboardSuggestion,
-  STARTUP_FILE_ACCEPT,
-  createClipboardSuggestionFromText,
-  getClipboardFilesFromEvent,
-  getClipboardSuggestion,
-  normalizeUrl,
-  startupFileExt,
-  startupFileKey,
-  startupLinkLabel,
-} from "@/components/app/create-agent-dialog-clipboard";
-import {
-  Popover,
-  PopoverContent,
-  PopoverTrigger,
-} from "@/components/ui/popover";
 import { type Agent } from "@/components/app/types";
 import { useClickOutside } from "@/hooks/use-click-outside";
 import { useRadixPopoverZFix } from "@/hooks/use-radix-popover-z-fix";
@@ -79,105 +57,6 @@ const FULL_ACCESS_PREFIX = "dispatch:fullAccess:";
 const AUTO_REVIEW_PREFIX = "dispatch:autoReview:";
 const BASE_BRANCH_PREFIX = "dispatch:baseBranch:";
 const CONTEXT_PROMPT_ID = "create-agent-context-prompt";
-const CONTEXT_LINK_INPUT_ID = "create-agent-context-link-input";
-const CONTEXT_LINK_ERROR_ID = "create-agent-context-link-error";
-
-function AddContextMenu({
-  onAddFile,
-  onAddLink,
-}: {
-  onAddFile: () => void;
-  onAddLink: () => void;
-}) {
-  return (
-    <div className="flex flex-col">
-      <button
-        type="button"
-        onClick={onAddFile}
-        className="flex w-full items-center gap-2 rounded px-2 py-1.5 text-sm text-foreground hover:bg-white/[0.1]"
-      >
-        <Upload className="h-3.5 w-3.5 text-muted-foreground" />
-        Add file
-      </button>
-      <button
-        type="button"
-        onClick={onAddLink}
-        className="flex w-full items-center gap-2 rounded px-2 py-1.5 text-sm text-foreground hover:bg-white/[0.1]"
-      >
-        <Link2 className="h-3.5 w-3.5 text-muted-foreground" />
-        Add link
-      </button>
-    </div>
-  );
-}
-
-function AddContextLinkForm({
-  value,
-  onChange,
-  onSubmit,
-  onBack,
-  isValid,
-}: {
-  value: string;
-  onChange: (next: string) => void;
-  onSubmit: () => void;
-  onBack: () => void;
-  isValid: boolean;
-}) {
-  return (
-    <div className="space-y-1.5 p-1">
-      <label
-        htmlFor={CONTEXT_LINK_INPUT_ID}
-        className="text-xs text-muted-foreground"
-      >
-        Link URL
-      </label>
-      <Input
-        id={CONTEXT_LINK_INPUT_ID}
-        autoFocus
-        value={value}
-        onChange={(event) => onChange(event.target.value)}
-        onKeyDown={(event) => {
-          if (event.key === "Enter") {
-            event.preventDefault();
-            if (value.trim().length > 0 && isValid) {
-              onSubmit();
-            }
-          }
-        }}
-        placeholder="https://..."
-        data-testid="create-agent-context-link-input"
-        aria-invalid={!isValid}
-        aria-describedby={!isValid ? CONTEXT_LINK_ERROR_ID : undefined}
-      />
-      {!isValid ? (
-        <p
-          id={CONTEXT_LINK_ERROR_ID}
-          className="text-xs text-status-blocked"
-          data-testid="create-agent-context-link-error"
-        >
-          Enter a valid `http:` or `https:` URL.
-        </p>
-      ) : null}
-      <div className="flex justify-between gap-2">
-        <Button type="button" variant="ghost" size="sm" onClick={onBack}>
-          <ChevronLeft className="mr-1 h-3 w-3" />
-          Back
-        </Button>
-        <Button
-          type="button"
-          variant="default"
-          size="sm"
-          onClick={onSubmit}
-          disabled={value.trim().length === 0 || !isValid}
-          data-testid="create-agent-context-link-add"
-        >
-          Add link
-        </Button>
-      </div>
-    </div>
-  );
-}
 
 function readStoredString(key: string): string {
   if (typeof window === "undefined") return "";
@@ -344,8 +223,6 @@ function CreateAgentDialogContent({
 }: Omit<CreateAgentDialogProps, "open">): JSX.Element {
   const [step, setStep] = useState<"config" | "context">("config");
   const promptTextareaRef = useRef<HTMLTextAreaElement>(null);
-  const startupFileInputRef = useRef<HTMLInputElement>(null);
-  const clipboardRequestIdRef = useRef(0);
   const [createName, setCreateName] = useState("");
   const [createType, setCreateType] = useState<AgentType>(() => {
     const preferred = initialAgentType ?? readLastUsedAgentType();
@@ -367,12 +244,8 @@ function CreateAgentDialogContent({
   const [initialPrompt, setInitialPrompt] = useState("");
   const [startupFiles, setStartupFiles] = useState<File[]>([]);
   const [startupLinks, setStartupLinks] = useState<string[]>([]);
-  const [linkDraft, setLinkDraft] = useState("");
-  const [checkingClipboard, setCheckingClipboard] = useState(false);
-  const [clipboardReadFeedback, setClipboardReadFeedback] = useState<
-    string | null
-  >(null);
   const [draggingFiles, setDraggingFiles] = useState(false);
+  const [contextDraftInvalid, setContextDraftInvalid] = useState(false);
   const [creating, setCreating] = useState(false);
   const [cwdHistory, setCwdHistory] = useState<string[]>(() =>
     readCwdHistory()
@@ -428,9 +301,6 @@ function CreateAgentDialogContent({
 
   useEffect(() => {
     if (step !== "context") {
-      clipboardRequestIdRef.current += 1;
-      setCheckingClipboard(false);
-      setClipboardReadFeedback(null);
       setDraggingFiles(false);
     }
   }, [step]);
@@ -480,37 +350,11 @@ function CreateAgentDialogContent({
     });
   }, []);
 
-  const applyClipboardSuggestion = useCallback(
-    (suggestion: ClipboardSuggestion) => {
-      switch (suggestion.kind) {
-        case "image":
-        case "file":
-          appendStartupFiles([suggestion.file]);
-          break;
-        case "url":
-          setStartupLinks((current) =>
-            current.includes(suggestion.url)
-              ? current
-              : [...current, suggestion.url]
-          );
-          break;
-        case "text": {
-          const suggestionText = suggestion.text;
-          // Append when the user has typed real content; treat whitespace-only
-          // as empty so we don't leave leading blank lines for someone who
-          // just clicked into the textarea before reading the clipboard.
-          setInitialPrompt((current) =>
-            current.trim().length === 0
-              ? suggestionText
-              : `${current.trimEnd()}\n\n${suggestionText}`
-          );
-          requestAnimationFrame(() => promptTextareaRef.current?.focus());
-          break;
-        }
-      }
-    },
-    [appendStartupFiles]
-  );
+  const handleAddLink = useCallback((url: string) => {
+    setStartupLinks((current) =>
+      current.includes(url) ? current : [...current, url]
+    );
+  }, []);
 
   const handleStartupPaste = useCallback(
     (event: ClipboardEvent<HTMLElement>) => {
@@ -523,25 +367,18 @@ function CreateAgentDialogContent({
       if (pastedFiles.length > 0) {
         event.preventDefault();
         appendStartupFiles(pastedFiles);
-        setClipboardReadFeedback(null);
         return;
       }
 
       const textSuggestion = createClipboardSuggestionFromText(
         event.clipboardData.getData("text/plain")
       );
-
-      // Plain text falls through to native paste so the prompt textarea
-      // gets the typed content. Only intercept URLs — the user almost
-      // always means "add this as a context link," not "paste this URL
-      // into the prompt as text."
       if (textSuggestion?.kind === "url") {
         event.preventDefault();
-        applyClipboardSuggestion(textSuggestion);
-        setClipboardReadFeedback(null);
+        handleAddLink(textSuggestion.url);
       }
     },
-    [appendStartupFiles, applyClipboardSuggestion]
+    [appendStartupFiles, handleAddLink]
   );
 
   const handleStartupDrop = useCallback(
@@ -551,15 +388,6 @@ function CreateAgentDialogContent({
       event.preventDefault();
       setDraggingFiles(false);
       appendStartupFiles(droppedFiles);
-    },
-    [appendStartupFiles]
-  );
-
-  const handleStartupFileChange = useCallback(
-    (event: ChangeEvent<HTMLInputElement>) => {
-      const selected = Array.from(event.target.files ?? []);
-      appendStartupFiles(selected);
-      event.target.value = "";
     },
     [appendStartupFiles]
   );
@@ -592,81 +420,16 @@ function CreateAgentDialogContent({
     );
   }, []);
 
+  const handleClipboardText = useCallback((text: string) => {
+    setInitialPrompt((current) =>
+      current.trim().length === 0 ? text : `${current.trimEnd()}\n\n${text}`
+    );
+    requestAnimationFrame(() => promptTextareaRef.current?.focus());
+  }, []);
+
   const enterContextStep = useCallback(() => {
     setStep("context");
-    setCheckingClipboard(false);
-    setClipboardReadFeedback(null);
   }, []);
-
-  const handleCheckClipboard = useCallback(() => {
-    setCheckingClipboard(true);
-    setClipboardReadFeedback(null);
-    const requestId = clipboardRequestIdRef.current + 1;
-    clipboardRequestIdRef.current = requestId;
-    void getClipboardSuggestion().then((result) => {
-      if (clipboardRequestIdRef.current !== requestId) return;
-      setCheckingClipboard(false);
-      if (result.suggestion) {
-        applyClipboardSuggestion(result.suggestion);
-        return;
-      }
-      setClipboardReadFeedback(
-        result.status === "blocked"
-          ? "Clipboard access was blocked. Paste into Instructions instead."
-          : result.status === "unsupported"
-            ? "Clipboard access isn't available here. Paste into Instructions instead."
-            : "Nothing readable found. Try pasting into Instructions instead."
-      );
-    });
-  }, [applyClipboardSuggestion]);
-
-  const normalizedLinkDraft = normalizeUrl(linkDraft);
-  const linkDraftIsValid =
-    linkDraft.trim().length === 0 || normalizedLinkDraft !== null;
-
-  const addStartupLink = useCallback(() => {
-    if (!normalizedLinkDraft) return false;
-    setStartupLinks((current) =>
-      current.includes(normalizedLinkDraft)
-        ? current
-        : [...current, normalizedLinkDraft]
-    );
-    setLinkDraft("");
-    return true;
-  }, [normalizedLinkDraft]);
-
-  const [addOpen, setAddOpen] = useState(false);
-  const [addMode, setAddMode] = useState<"menu" | "link">("menu");
-
-  const handleAddOpenChange = useCallback((next: boolean) => {
-    setAddOpen(next);
-    if (!next) {
-      setAddMode("menu");
-      setLinkDraft("");
-    }
-  }, []);
-
-  const handleAddFileFromMenu = useCallback(() => {
-    setAddOpen(false);
-    setAddMode("menu");
-    // Defer the file picker click so the popover unmounts cleanly first.
-    requestAnimationFrame(() => startupFileInputRef.current?.click());
-  }, []);
-
-  const handleAddLinkFromMenu = useCallback(() => {
-    setAddMode("link");
-  }, []);
-
-  const handleAddLinkBack = useCallback(() => {
-    setLinkDraft("");
-    setAddMode("menu");
-  }, []);
-
-  const handleAddLinkSubmit = useCallback(() => {
-    if (!addStartupLink()) return;
-    setAddMode("menu");
-    setAddOpen(false);
-  }, [addStartupLink]);
 
   useRadixPopoverZFix();
 
@@ -675,7 +438,6 @@ function CreateAgentDialogContent({
       event.preventDefault();
       const cwd = createCwd.trim();
       if (!cwd) return;
-      if (step === "context" && !linkDraftIsValid) return;
 
       setCreating(true);
       try {
@@ -754,7 +516,6 @@ function CreateAgentDialogContent({
       createUseWorktree,
       createWorktreeBranch,
       initialPrompt,
-      linkDraftIsValid,
       onCreated,
       startupFiles,
       startupLinks,
@@ -1145,37 +906,6 @@ function CreateAgentDialogContent({
               )}
             >
               <div className="space-y-3">
-                <div
-                  className="rounded-lg border border-dashed border-white/[0.12] bg-white/[0.03] px-3 py-3"
-                  data-testid="create-agent-context-clipboard-check"
-                >
-                  <Button
-                    type="button"
-                    variant="default"
-                    className="min-h-11 w-full justify-center px-3"
-                    onClick={handleCheckClipboard}
-                    disabled={checkingClipboard}
-                    data-testid="create-agent-context-clipboard-check-action"
-                  >
-                    {checkingClipboard ? (
-                      <ActivityBars size={16} className="mr-1.5" />
-                    ) : (
-                      <Clipboard className="mr-1.5 h-3.5 w-3.5" />
-                    )}
-                    Read clipboard
-                  </Button>
-                  {clipboardReadFeedback ? (
-                    <p
-                      className="mt-2 text-xs text-muted-foreground"
-                      role="status"
-                      aria-live="polite"
-                      data-testid="create-agent-context-clipboard-feedback"
-                    >
-                      {clipboardReadFeedback}
-                    </p>
-                  ) : null}
-                </div>
-
                 <div className="space-y-1">
                   <label
                     htmlFor={CONTEXT_PROMPT_ID}
@@ -1199,205 +929,19 @@ function CreateAgentDialogContent({
                   />
                 </div>
 
-                <div
-                  className={cn(
-                    "space-y-3 rounded-md border bg-muted/20 px-3 py-3 transition-colors",
-                    draggingFiles
-                      ? "border-status-done/35 bg-status-done/8 ring-1 ring-inset ring-status-done/30"
-                      : "border-border/70"
-                  )}
-                >
-                  <div>
-                    <div className="text-sm font-medium text-foreground">
-                      Context
-                    </div>
-                    <p className="text-xs text-muted-foreground">
-                      Attach files or links to give the agent context.
-                    </p>
-                  </div>
-                  <input
-                    ref={startupFileInputRef}
-                    type="file"
-                    multiple
-                    accept={STARTUP_FILE_ACCEPT}
-                    className="hidden"
-                    onChange={handleStartupFileChange}
-                    data-testid="create-agent-context-files-input"
-                  />
-                  {startupFiles.length === 0 && startupLinks.length === 0 ? (
-                    <Popover open={addOpen} onOpenChange={handleAddOpenChange}>
-                      <PopoverTrigger asChild>
-                        <button
-                          type="button"
-                          data-testid="create-agent-context-files-button"
-                          className={cn(
-                            "flex w-full flex-col items-center justify-center gap-1.5 rounded-md border border-dashed px-4 py-6 text-sm transition-colors",
-                            draggingFiles
-                              ? "border-status-done bg-status-done/10 text-foreground"
-                              : "border-border/70 bg-background/40 text-muted-foreground hover:border-border hover:bg-background/70 hover:text-foreground"
-                          )}
-                        >
-                          {draggingFiles ? (
-                            <>
-                              <Upload className="h-5 w-5" />
-                              <span>Drop file to add</span>
-                            </>
-                          ) : (
-                            <>
-                              <Plus className="h-5 w-5" />
-                              <span>Add files or links</span>
-                              <span className="text-[11px] text-muted-foreground/80">
-                                Click to choose, or drop a file here
-                              </span>
-                            </>
-                          )}
-                        </button>
-                      </PopoverTrigger>
-                      <PopoverContent
-                        align="start"
-                        className="w-72 p-1"
-                        forceMount
-                      >
-                        {addMode === "menu" ? (
-                          <AddContextMenu
-                            onAddFile={handleAddFileFromMenu}
-                            onAddLink={handleAddLinkFromMenu}
-                          />
-                        ) : (
-                          <AddContextLinkForm
-                            value={linkDraft}
-                            onChange={setLinkDraft}
-                            onSubmit={handleAddLinkSubmit}
-                            onBack={handleAddLinkBack}
-                            isValid={linkDraftIsValid}
-                          />
-                        )}
-                      </PopoverContent>
-                    </Popover>
-                  ) : (
-                    <div className="flex flex-wrap items-start gap-3">
-                      {startupFiles.map((file) => {
-                        const key = startupFileKey(file);
-                        const preview = startupFilePreviewsRef.current.get(key);
-                        return (
-                          <div
-                            key={key}
-                            className="group flex w-12 flex-col gap-0.5"
-                          >
-                            <div className="relative h-12 w-12 overflow-hidden rounded-md border border-border/70 bg-muted/40">
-                              {preview ? (
-                                <img
-                                  src={preview}
-                                  alt=""
-                                  className="h-full w-full object-cover"
-                                />
-                              ) : (
-                                <div className="flex h-full w-full flex-col items-center justify-center text-muted-foreground">
-                                  <FileText className="h-3.5 w-3.5" />
-                                  <span className="text-[8px] font-medium tracking-wide">
-                                    {startupFileExt(file.name)}
-                                  </span>
-                                </div>
-                              )}
-                              <button
-                                type="button"
-                                className="absolute -right-2 -top-2 flex h-10 w-10 items-start justify-end rounded-full p-2 text-muted-foreground transition-opacity hover:text-foreground focus:opacity-100 group-hover:opacity-100"
-                                onClick={() => handleRemoveStartupFile(file)}
-                                aria-label={`Remove ${file.name}`}
-                              >
-                                <span className="rounded-full border border-border/70 bg-background p-0.5">
-                                  <X className="h-2.5 w-2.5" />
-                                </span>
-                              </button>
-                            </div>
-                            <span
-                              className="w-full truncate text-[8px] leading-tight text-muted-foreground"
-                              title={file.name}
-                            >
-                              {file.name}
-                            </span>
-                          </div>
-                        );
-                      })}
-                      {startupLinks.map((link) => {
-                        const { host, rest } = startupLinkLabel(link);
-                        return (
-                          <div
-                            key={link}
-                            className="group relative flex h-12 max-w-[180px] flex-col justify-center gap-0.5 rounded-md border border-border/70 bg-muted/40 px-2 pr-7 leading-tight"
-                            title={link}
-                          >
-                            <div className="flex items-center gap-1 text-[10px] text-foreground">
-                              <Link2 className="h-2.5 w-2.5 shrink-0 text-muted-foreground" />
-                              <span className="truncate font-medium">
-                                {host}
-                              </span>
-                            </div>
-                            {rest ? (
-                              <span className="truncate text-[9px] text-muted-foreground">
-                                {rest}
-                              </span>
-                            ) : null}
-                            <button
-                              type="button"
-                              className="absolute -right-2 -top-2 flex h-10 w-10 items-start justify-end rounded-full p-2 text-muted-foreground transition-opacity hover:text-foreground focus:opacity-100 group-hover:opacity-100"
-                              onClick={() => handleRemoveStartupLink(link)}
-                              aria-label={`Remove ${link}`}
-                            >
-                              <span className="rounded-full border border-border/70 bg-background p-0.5">
-                                <X className="h-2.5 w-2.5" />
-                              </span>
-                            </button>
-                          </div>
-                        );
-                      })}
-                      <Popover
-                        open={addOpen}
-                        onOpenChange={handleAddOpenChange}
-                      >
-                        <PopoverTrigger asChild>
-                          <button
-                            type="button"
-                            data-testid="create-agent-context-files-button"
-                            aria-label="Add context"
-                            className={cn(
-                              "flex h-12 w-12 items-center justify-center rounded-md border border-dashed transition-colors",
-                              draggingFiles
-                                ? "border-status-done bg-status-done/10 text-foreground"
-                                : "border-border/70 bg-background/40 text-muted-foreground hover:border-border hover:bg-background/70 hover:text-foreground"
-                            )}
-                          >
-                            {draggingFiles ? (
-                              <Upload className="h-4 w-4" />
-                            ) : (
-                              <Plus className="h-4 w-4" />
-                            )}
-                          </button>
-                        </PopoverTrigger>
-                        <PopoverContent
-                          align="start"
-                          className="w-72 p-1"
-                          forceMount
-                        >
-                          {addMode === "menu" ? (
-                            <AddContextMenu
-                              onAddFile={handleAddFileFromMenu}
-                              onAddLink={handleAddLinkFromMenu}
-                            />
-                          ) : (
-                            <AddContextLinkForm
-                              value={linkDraft}
-                              onChange={setLinkDraft}
-                              onSubmit={handleAddLinkSubmit}
-                              onBack={handleAddLinkBack}
-                              isValid={linkDraftIsValid}
-                            />
-                          )}
-                        </PopoverContent>
-                      </Popover>
-                    </div>
-                  )}
-                </div>
+                <ContextPicker
+                  files={startupFiles}
+                  links={startupLinks}
+                  draggingFiles={draggingFiles}
+                  filePreviewsRef={startupFilePreviewsRef}
+                  onAppendFiles={appendStartupFiles}
+                  onRemoveFile={handleRemoveStartupFile}
+                  onAddLink={handleAddLink}
+                  onRemoveLink={handleRemoveStartupLink}
+                  onClipboardText={handleClipboardText}
+                  onDraftInvalid={setContextDraftInvalid}
+                  testIdPrefix="create-agent-context"
+                />
               </div>
             </div>
 
@@ -1428,7 +972,7 @@ function CreateAgentDialogContent({
                   variant="primary"
                   tabIndex={0}
                   className="min-h-11 px-3"
-                  disabled={creating || !linkDraftIsValid}
+                  disabled={creating || contextDraftInvalid}
                   data-testid="create-agent-context-submit"
                 >
                   {creating ? (
